add: running a Script with a main method error
when working with a script im groovy you cannot use some class with a main
method because the body of the script internally is wrapped by anothe class
which heritages from script class. and this superclass has its own public main
method. So it cannot be runned. However it can be runned, if we only separate
definition of classes in multiple files or remove the main method and treat is
a a script