fix: main method in the first class defined in a script
see more at https://stackoverflow.com/questions/23717406/groovy-main-method-placement

"You can get a clear picture when you inspect AST in groovy console"
The Abstract Syntax Tree (AST) in the Groovy console can be used to inspect the structure and elements of a Groovy script. The AST provides a tree-like representation of the code, with each node in the tree representing a different element of the script, such as a class, method, or variable. Inspecting the AST can be useful for understanding the structure of the code and for debugging or troubleshooting issues in the script.